feat(ui): show child folders in three-column note list
