--- .github/workflows/tmp-three-column-folder-contents-v4.yml
@@ -1,63 +1,0 @@
-name: Finalize three-column folder contents patch
-
-on:
-  push:
-    branches:
-      - feat/v1.4.6-three-column-folder-contents
-
-permissions:
-  contents: write
-
-jobs:
-  apply:
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: feat/v1.4.6-three-column-folder-contents
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-          cache-dependency-path: frontend/package-lock.json
-
-      - name: Repair staged patch targets
-        run: |
-          python scripts/tmp-fix-three-column-patch.py
-          python scripts/tmp-fix-three-column-patch-v2.py
-
-      - name: Apply patch
-        run: python scripts/tmp-apply-three-column-folder-contents.py
-
-      - name: Install frontend dependencies
-        working-directory: frontend
-        run: npm ci
-
-      - name: Run focused tests
-        working-directory: frontend
-        run: npm run test:run -- src/lib/__tests__/threeColumnFolderContents.test.ts src/lib/__tests__/knowledgeTreeSidebarContract.test.ts src/lib/__tests__/noteWorkspaceLayout.test.ts
-
-      - name: Type-check frontend
-        working-directory: frontend
-        run: npx tsc -b --pretty false
-
-      - name: Commit verified implementation
-        shell: bash
-        run: |
-          git rm .github/workflows/tmp-three-column-folder-contents.yml
-          git rm .github/workflows/tmp-three-column-folder-contents-v2.yml
-          git rm .github/workflows/tmp-three-column-folder-contents-v3.yml
-          git rm .github/workflows/tmp-three-column-folder-contents-v4.yml
-          git rm scripts/tmp-apply-three-column-folder-contents.py
-          git rm scripts/tmp-fix-three-column-patch.py
-          git rm scripts/tmp-fix-three-column-patch-v2.py
-          git config user.name "Drug"
-          git config user.email "cropflre2333@gmail.com"
-          git add frontend/src/lib/threeColumnFolderContents.ts
-          git add frontend/src/lib/__tests__/threeColumnFolderContents.test.ts
-          git add frontend/src/components/NoteList.tsx
-          git add frontend/src/components/KnowledgeTreePanel.tsx
-          git commit -m "feat(ui): show child folders in three-column note list"
-          git push origin HEAD:feat/v1.4.6-three-column-folder-contents
